Upload Data to Index. Push props whith Route.
