+++ src/App.css
@@ -11,8 +11,9 @@
 
 .app-wrapper-content {
     grid-area: c;
-    background-color: navajowhite;
-    width: 1100px;
+    background-color: mediumaquamarine;
+    width: 1050px;
+
  }
 .app-wrapper {
     font-family: Montserrat, sans-serif;
@@ -22,6 +23,7 @@
     line-height: 1.8;
     padding: 0;
     margin: 0;
+    padding-top: 5px;
     }
