+++ src/components/Dialogs/Dialogs.jsx
@@ -1,48 +1,14 @@
 import React from "react";
 import s from "./Dialogs.module.css";
 import {NavLink} from "react-router-dom";
-
-const DialogItem = (props) => {
-
-    let path = "/dialogs/" + props.id;
-
-    return <div className={s.dialog + ' ' + s.active}>
-
-        <NavLink to={path}>{props.name}</NavLink>
-    </div>
-}
-
-const Message = (props) => {
-    return <div className={s.dialog}>{props.message}</div>
-}
+import DialogItem from "./DialogItem/dialogItem";
+import Message from "./Message/message";
 
 const Dialogs = (props) => {
-    let dialogs = [
-        {id: 1, name: 'Serhio'},
-        {id: 2, name: 'Olyalya'},
-        {id: 3, name: 'Dashunya'},
-        {id: 4, name: 'Yanki'},
-        {id: 5, name: 'Pupkin'},
-        {id: 6, name: 'Artem'},
-        {id: 7, name: 'Roman'}
-    ]
-
-    let messages = [
-        {id: 1, message: 'Hallo!!!'},
-        {id: 2, message: 'I am very glad to see you!!!'},
-        {id: 3, message: 'It is so fun'},
-        {id: 4, message: 'Yo'},
-        {id: 5, message: 'yo'},
-        {id: 6, message: 'yoooo'},
-        {id: 7, message: 'yoooooouuu'}
-    ]
-
-    let dialogsElements = dialogs
+    let dialogsElements = props.dialogs
         .map(d => <DialogItem name={d.name} id={d.id}/>)
-
-    let messagesElements = messages
+    let messagesElements = props.messages
         .map(m => <Message message={m.message}/>)
-
     return (
         <div className={s.dialogs}>
             <div className={s.dialogsItems}>
