+++ src/components/Header/Header.module.css
@@ -2,7 +2,8 @@
     grid-area: h;
     background-color: whitesmoke;
     padding-left: 20px;
-}
+    border-radius: 5px;
+   }
 
 .header img {
     width: 100px;
