+++ src/components/Navbar/Navbar.module.css
@@ -2,8 +2,10 @@
 
     grid-area: n;
     width: 100px;
+    height: 400px;
     background-color: whitesmoke;
     padding: 20px;
+    border-radius: 5px;
 }
 
 .item a {
