+++ src/components/Profile/Profile.module.css
@@ -1,7 +1,7 @@
 
 .wallpepper {
     width: 1050px;
-    height: 250px;
+    height: 400px;
     background-color: navajowhite;
 }
